Solutions/Userreportedphishingv2/temp/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
azure-functions==1.20.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorclass==2.2.2
compressed_rtf==1.0.6
cryptography==43.0.0
easygui==0.98.3
ebcdic==1.1.1
extract-msg==0.48.7
filelock==3.15.4
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
lark==1.1.9
msoffcrypto-tool==5.4.1
multidict==6.0.5
olefile==0.47
oletools==0.60.2
pcodedmp==1.2.6
pycparser==2.22
pyparsing==3.1.2
red-black-tree-mod==1.20
requests==2.31.0
requests-file==2.1.0
RTFDE==0.1.2
scapy==2.5.0
sniffio==1.3.1
soupsieve==2.5
tldextract==5.1.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
win_unicode_console==0.5
yarl==1.9.4